Check for interrupt status in recursive sort call
Summary:
We have observed sort tasks continue to run even after the query is killed.
Looking at the sort code, it looks like currently the sort is recursive call and
doesn't check for the interrupt status of the thread. When a query is killed all
its tasks are interrupted. It is up the code to check for interrupt signal and take
action. In our case, we just need to throw an exception and continue

Reviewers: dharakk

Reviewed By: dharakk

Subscribers: jenkins

Maniphest Tasks: T2270079

Differential Revision: https://code.uberinternal.com/D4136475