Provide a config to disable resource groups entirely
Summary:
this is obviously disabled by default ! The reason we need this is
that even with the InternalResourceGroup sometimes the query can spend upto 300
ms just to enter the "start" call. I suspect the lock in the
InternalResourceGroup is to blame (under load).

So for neutrino we wanted a way to simply bypass this entirely.

Super open to suggestions to how to do this cleanly without changing a lot of code :-)

I thought of repurposing or creating a new NoopResourceGroup but that one is more geared for the tasks. Perhaps I could introduce a ReallyNoopResourceGroupForCoordinator to solve my use case.

Reviewers: bhasudha, ujwala

Reviewed By: bhasudha, ujwala

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D4338525