[Neutrino] Have a way to provide default values for X-Presto-User and X-Presto-Wait-For-Entire-Response-Ms headers
Summary:
Reduce the number of headers etc that must be set by the
callers and hence the flakyness. With muttley, RPC-Caller suffices for
X-Presto-User. And we can have a default timeout, that can be optionally
overridden.

The new interface is backwards compatible.

Test Plan:
Test with specifying and not
specifying these headers, with and without the config file.

Reviewers: venkik, lu.niu, curt, bhasudha

Reviewed By: curt

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D3083511