[Perf logging] Plumbing the sessionLogEntries to the query-event
Summary:
The current way to get the session log entries from the UI is
cumbersome particularly because they get garbage collected. Thus, I
would like to expose them out to the query event listener just like
their sibling operatorSummaries.

I will also have neutrino pingless collect this logging info. This does
not have any impact on photon's query logging because this field is
absent in the hp-presto-query-event topic and thus would be skipped
over. (not to mention that we rarely expect to enable the
X-Presto-Query-Logging-Size header to record this information in the
first place)

Reviewers: venkik, curt, bhasudha

Reviewed By: venkik

Differential Revision: https://code.uberinternal.com/D3731711