[Neutrino][Easy] Introduce a default configurable catalog
Summary:
In the quest to continue to make the neutrino /v1/statement endpoint
be as natural as possible, also allow the catalog field (in addition to the
user as done before) to be configurable.

Thus now users don't need to specify the X-Presto-Catalog=rta field.

Test Plan: Test with neutrino staging with curl and the presto-client. Also test by running presto-main (non neutrino) codepath

Reviewers: #neutrino, bhasudha

Reviewed By: #neutrino, bhasudha

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D4092897