[Uber Event Listener] Record the nonScanBlockTime in the query log kafka topic
Summary:
This metric was already being recorded to m3 and I would like to push it out to kafka also, so that we can build ELK histograms on it. Unfortunately, ELK cannot do histograms on expressions (ie cannot histogram elapsedTimeMs - scanBlockTimeMs).

Once this diff is landed, I will push this new field out to the Neutrino heatpipe topic.

This extra field won't have any impact on the existing Presto/Photon heatpipe topic if it is not added to the schema.

Test Plan: Live

Reviewers: bhasudha, ujwala

Reviewed By: bhasudha

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D4384447