Adding instrumentation framework to log query lifecycle events
This framework provides ability to log events with timestamp during query's
execution lifecycle. The basic usage is:
session.getSessionLogger().log(() -> String.format("some message")).

These messages can be logged at the connector
(using the ConnectorSession) or at the coordinator or tasks (that have access
to the Session). The log messages are shipped as a part of the regular
QueryInfo from all the tasks and show up in the final json: there is one
"sessionLogEntries" for the QueryInfo and one per each of the tasks in the
query.

This is disabled by default and can be enabled by using the header
X-Presto-Query-Logging-Size on the query.

Added three examples of usage loggin planning compeletion and plan
distribution events. sessionLogEntries for a sample query is as following:

"sessionLogEntries" : [ {
  "message" : "starting query",
  "nanos" : 659651498121659,
  "threadName" : "Query-20191031_045412_00002_7p8ay-390"
}, {
  "message" : "done planning",
  "nanos" : 659651538419849,
  "threadName" : "Query-20191031_045412_00002_7p8ay-390"
}, {
  "message" : "done plan distribution",
  "nanos" : 659651540308691,
  "threadName" : "Query-20191031_045412_00002_7p8ay-390"
} ]