Use the Parquet column name from the schema in file
Summary:
Currently we try to recreate the subfield type using column/subfield name given
by the Presto planner. This causes issues later on when we try to look up ColumnIO
using `ColumnIOFactory.getColumnIO(MessageType requestedSchema, MessageType fileSchema)`
which is case-sensitive column name look up code.

This issue only reproes when Parquet decryption is enabled as there are differences in how
the schema from the file read.
When decryption is disabled we use `MetadataReader.readFooter(FSDataInputStream inputStream, Path file, long fileSize)`
which normalizes the column names to lowercase.
When decryption is enabled we use `MetadataReader.readFooter(Path file,
                                                   long fileSize,
                                                   ParquetReadOptions options,
                                                   SeekableInputStream inputStream,
                                                   ParquetMetadataConverter converter,
                                                   FileDecryptionProperties fileDecryptionProperties,
                                                   InternalFileDecryptor fileDecryptor)`
which doesn't normalize the column names.

We should fix the MetadataReader to have same path irrespective of descryption is enabled or not.
This will ensure future opensource changes in Parquet that assume case normalized column names in
schemar read from the file.

This fix is a temporary fix that reverts to 217 behavior for nested column pushdown.

Reviewers: shangx, pavi, vivekb

Reviewed By: shangx, vivekb

Maniphest Tasks: T5830635

Differential Revision: https://code.uberinternal.com/D4312847