Fix issue that is causing the tests to not run in Jenkins
Summary:
HDFS observer property is peculiar property that is not used like the other properties.
We need to manually add them in the test connector session if it is not present.
Otherwise this is causing the tests to fail in test VM start up and no tests are run.

Reviewers: dharakk

Reviewed By: dharakk

Maniphest Tasks: T5836053

Differential Revision: https://code.uberinternal.com/D4314823