Handle pruned columns in a pushed down pinot table scan
Summary:
As an example, consider this query:
```
select count(*) from baseballstats group by teamid limit 10
```

In this query, the pinot connector sees the plan:
Project (count) -> Limit 10 -> Aggregation (teamid, count(*)) -> Scan baseballstats.

Since projections cannot currently be pushed on top of aggregations, so only the plan until the limit is pushed down. The resulting plan looks like: Project(count) -> Pushed table scan node with PQL: "select count(*) from baseballstats group by teamid". Note that the pushed table scan will emit "teamid" and "count(*)".

PruneUnreferencedOptimizer runs next and removes the "teamid" from the scan node. This creates a problem: we are emitting an extra column teamid in the PinotBrokerPageSource which the scan does not need.

This PR attempts to handle this scenario: both in the broker and segment codepaths. It does so with a lot of plumbing: It remembers the original column handles when the PQL was generated in the PinotTableHandle. These are passed eventually to the page sources via the PinotSplit.

The page source then checks which columns can be ignored or have been re-ordered. The broker page source already had some support for this (necessitated by the _hiddenColumnSet_ for handling group-by without aggregations) and I have merely generalized this support. It was also trivial to add this support to the segment page source.

Added unit tests for these conditions in the pinot broker and segment page sources.

Depends on: D4391447

Reviewers: ujwala, bhasudha

Reviewed By: ujwala

Subscribers: O4263 subscribe to presto changes

Differential Revision: https://code.uberinternal.com/D4391577