[Pinot Connector] Fixing a bug in hidden column handling
Summary:
In the diff D4559101 we removed the hidden column from output variables but messed up by also removing from the table handle. This diff fixes it fundamentally. Flow goes like this

aggregation node has group by column but no aggregation -> we add `count(*)` as hidden column and add it to expected column from pql response and remember it for later usage in pagesource -> at the end of plan optimization we remove hiddn column from output variables and assignment of the `TableScanNode` as they are really not going to be in the output -> when we get response from PQL we have remembered that we get two columns and we intentionally now ignore the hiddne column and return what is expected in scan node's output.

Reviewers: venkik, curt, lxning, ujwala

Reviewed By: venkik, curt, ujwala

Subscribers: O4263 subscribe to presto changes

Differential Revision: https://code.uberinternal.com/D4569067