Fix bug in Sample Table Name replacement
Summary: Yesterday, I noticed that the sample table name replacement was not working when the parent table appeared multiple times in the original. It replaces the first occurance of the table but not the subsequent ones. Fixed the issue and make it a failure in the code if that happens again. Also, added a test case that will catch that condition.

Test Plan: Added a unit test to catch the bug being fixed. Also verified that the query https://datacentral.uberinternal.com/presto/queries/20200712_031701_02640_25xyf now gets handled properly. Both the occurences are replaced https://code.uberinternal.com/P236277 instead of the single occurance earlier https://code.uberinternal.com/P236231

Reviewers: venkik

Reviewed By: venkik

Subscribers: O4263 subscribe to presto changes

Maniphest Tasks: T6102659

Differential Revision: https://code.uberinternal.com/D4685317