Add support in parquet reader for reading TIMESTAMP_MICROS type.
Summary:
Right now presto always assumes TIMESTAMP_MILLIS as the OriginalType when int64 is used to
represent timestamp type in the schema. This causes two issues.

1. When we use createCryptoParquetPageSource we use parquet's own
conversion utils and we end up interpreting micro seconds as milli seconds and we show a date
in the distant future.
2. When we use createParquetPageSource we use our own convertor and this does not have a check
for TIMESTAMP_MILLIS and this can fail the query

Fix: In this change I am adding two ValueDecoders one for plain and one for RLE compression format. I
am also adding support for the v1 reader. All of these are in 3 new classes so that while creation we
instantiate the class based on the OriginalType. One Alternate approach is to do this check inside
LongValueDecoders but that would make this check inside the most critical path and can affect query
performance. The fix simply divides the micros seconds by 1000 to get the milliseconds and this should
be ok because in presto we anyway operate at the millisecond granularity.

For testing the ValueDecoders have their own unit test. Apart from that have also added a new test in
ParquetTester to check if timestamp stored as int64 with OriginalType TIMESTAMP_MICROS works.

As part of this change have also added two new SESSIONs in the full test mode so that we now get more code
coverage for the ParquetReader test ( we were not testing creating page source without encryption,
using non optimized reader and so on. )

Note: We use hive parquet writer and schema definition as defined in the presto-hive package
which is a shaded jar with a old version of hive. This packs an old version of parquet-mr that does not
have the TIMESTAMP_MICROS type. However, in the read path, we use the independent parquet dependency
which is more recent and so we use the version with the new enum. If we upgrade all the tests
to use the new version of parquet, we would not be testing parquet version writeen
by the packaged hive writer. We would need a version of the test that would work with the new definitions.
For now I have done this only for the timestamp type to test TIMESTAMP_MICROS. As part of seperate change
we can make sure all the test are tested with latest parquet writer. This change uses ExampleParquetWriter
which is supposed to be used only for demo or test purposes.

Test Plan: test in staging and local docker setup and added unit tests

Reviewers: venkik

Reviewed By: venkik

Subscribers: O4263 subscribe to presto changes

Differential Revision: https://code.uberinternal.com/D4772161