[Pinot Connector] Creating a custom key for broker cache
Summary:
Since the Neutrino 230 upgrade, we are seeing high latencies for Pinot Get calls. This latency was stemming mainly from the `/instances` calls to Pinot.
The root cause is that the broker values are not being cached and we were making calls for each query.

There were two problem with the caching key class `PinotTableHandle`:
1) The `GeneratedPql` class within does not implement `equals` and `hashCode` method so each time the hash value was different for same table handle
2) Even if we fix this, because of having pql part of the key, different queries on the same table will end up making the get call again rather than reusing the cached value

Because of these, it is not a good idea to use `PinotTableHandle` as the caching key. So I am creating a new class for key with equals and hashcode properly implemented

Test Plan: Checked locally that values are being cached and we are not making get calls for each query.

Reviewers: ujwala, yupeng, #ldap_rta-group

Reviewed By: yupeng, #ldap_rta-group

Subscribers: qiaochu, O4263 subscribe to presto changes

Differential Revision: https://code.uberinternal.com/D4927957