upgrade hadoop-apache2 to 2.8.2-uber-15
Summary:
Upgrade hadoop-apache2 to 2.8.2-uber-15 which uses 2.8.2-uber-135.

HDFS is planning to upgrade namenode/observer machine to a more powerful SKU next week.

To avoid restarting the whole cluster and redeploying configuration to all clients, we
will rename the existing machine and apply the original host name to the new host. The
current version of Hadoop client cannot handle this case: it caches IP addresses of name
nodes and will not be able to locate the new name node.

Reviewers: venkik

Reviewed By: venkik

Subscribers: O4263 subscribe to presto changes

JIRA Issues: HDFS-253

Differential Revision: https://code.uberinternal.com/D4969449