Add null masking to encryption
Summary:
The use case is like this. When a user gets access denied for a to encrypt table, Presto can indicate in the error message that if null is acceptable for the column xxx to your query, please enable 'read_masked_value_enabled'.

The whole code changes are behind a flag with default value false.

Test Plan: Test a table it in local intelliJ with SSH tunneling to HMS.

Reviewers: #ldap_presto-core, venkik, pavi

Reviewed By: #ldap_presto-core, venkik, pavi

Subscribers: O4263 subscribe to presto changes

Differential Revision: https://code.uberinternal.com/D5032245