Fix the mbean classnames in Odin metric collection script
Summary:
Odin metric collection script (which is based on Neutrino)
is bit rigid when looking up the metric for a given mbean classname.
It expects it to match exactly like what the process exports it.
In this case the order of "name", "type" are reversed, which
causes the lookup code to ignore the metric.

Odin metric script works different from Ubermon. In Ubermon we make
individual JMX calls, in Odin we export all JMX metrics from process
and look for the ones we need.

Reviewers: #ldap_presto-core, hitarth, vineeth

Reviewed By: #ldap_presto-core, hitarth, vineeth

Subscribers: O4263 subscribe to presto changes

Differential Revision: https://code.uberinternal.com/D5162627